Split `plugin-api.ts` into two files and move into `compat`
